+++ DuplicateError.js
@@ -1,0 +1,16 @@
+/*!
+ * Copyright (c) 2018 Digital Bazaar, Inc. All rights reserved.
+ */
+'use strict';
+
+export default class DuplicateError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = 'DuplicateError';
+    if(typeof Error.captureStackTrace === 'function') {
+      Error.captureStackTrace(this, this.constructor);
+    } else {
+      this.stack = (new Error(message)).stack;
+    }
+  }
+}