+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # bedrock-web-store ChangeLog
 
+### Added
+- Add `DuplicateError` name to duplicate errors.
+
 ## 1.0.0 - 2018-06-06
 
 ### Changed